requirements.txt
python-telegram-bot==20.6
supabase==1.0.3
geopandas==1.2.2
shapely==2.1.1
pandas==2.1.1
numpy==1.26.5
flask==2.3.4
gunicorn==21.2.0
fastapi==0.111.1
uvicorn==0.25.1
